Added a working test for dictionary routines. It works!
- comp_dict.c / comp_dict.h
Dictionary item structure's changed. The value associated to a key is
not a string anymore. Now it is an integer.

- util.h
Created a useful macro to get the size of a given array: ARRAY_SIZE().

- CMakeLists.txt
Lines related to generation of test binaries were added as comments.
Added a line that makes it possible to generate binaries with debug
symbols, so we can use gdb.